+++ src/cli/commands/public-cmds/link-cmd.js
@@ -2,6 +2,7 @@
 import chalk from 'chalk';
 import Command from '../../command';
 import { link } from '../../../api/consumer';
+import linkTemplate from '../../templates/link-template';
 
 export default class Create extends Command {
   name = 'link';
@@ -16,22 +17,6 @@ export default class Create extends Command {
   }
 
   report(results: Array<{ id: string, bound: ?Object }>): string {
-    const reportComponents = results
-      .map((result) => {
-        const bounds = result.bound
-          .filter(bound => bound.from && bound.to)
-          .map(bound => `\t\tfrom: ${bound.from}, to: ${bound.to}`)
-          .join('\n');
-        if (!bounds.length) {
-          const reason = result.id.scope ? 'is a nested dependency' : 'was not exported yet';
-          return chalk.cyan(`\t${result.id}:\n\t\tnothing to link because the component ${reason}`);
-        }
-        return chalk.cyan(`\t${result.id}:\n ${bounds}`);
-      })
-      .join('\n');
-
-    const reportTitle = chalk.underline(`found ${chalk.bold(results.length)} components\n`);
-
-    return reportTitle + reportComponents;
+    return linkTemplate(results);
   }
 }