+++ e2e/commands/add.e2e.js
@@ -369,7 +369,7 @@ describe('bit add command', function () {
       expect(errorMessage).to.have.string("error: option `-n, --namespace <namespace>' argument missing");
     });
     it('Define dynamic main file ', () => {
-      const mainFileOs = path.normalize('{PARENT_FOLDER}/{PARENT_FOLDER}.js');
+      const mainFileOs = path.normalize('{PARENT}/{PARENT}.js');
       helper.createComponent('bar', 'bar.js');
       helper.createComponent('bar', 'foo1.js');
       helper.addComponentWithOptions('bar', { m: mainFileOs, n: 'test' });
@@ -477,7 +477,7 @@ describe('bit add command', function () {
       helper.addComponentWithOptions('bar/', {
         i: 'bar/foo',
         m: path.normalize('bar/foo.js'),
-        t: 'test/{PARENT_FOLDER}/{FILE_NAME}.spec.js,test/{FILE_NAME}.spec.js'
+        t: 'test/{PARENT}/{FILE_NAME}.spec.js,test/{FILE_NAME}.spec.js'
       });
       const bitMap = helper.readBitMap();
       const files = bitMap['bar/foo'].files;
@@ -497,7 +497,7 @@ describe('bit add command', function () {
       helper.addComponentWithOptions('bar/*.js', {
         i: 'bar/foo',
         m: path.normalize('bar/foo.js'),
-        t: 'test/{PARENT_FOLDER}/{FILE_NAME}.spec.js,bar/foo.spec.js,test/{FILE_NAME}.spec.js'
+        t: 'test/{PARENT}/{FILE_NAME}.spec.js,bar/foo.spec.js,test/{FILE_NAME}.spec.js'
       });
       const bitMap = helper.readBitMap();
       const files = bitMap['bar/foo'].files;
@@ -518,7 +518,7 @@ describe('bit add command', function () {
       helper.addComponentWithOptions('bar/', {
         i: 'bar/foo',
         m: path.normalize('bar/foo.js'),
-        t: 'test/{PARENT_FOLDER}/{FILE_NAME}.spec.js,test/*.spec.js'
+        t: 'test/{PARENT}/{FILE_NAME}.spec.js,test/*.spec.js'
       });
       const bitMap = helper.readBitMap();
       const files = bitMap['bar/foo'].files;
@@ -553,7 +553,7 @@ describe('bit add command', function () {
       helper.addComponentWithOptions('bar/', {
         i: 'bar/foo',
         m: path.normalize('bar/foo.js'),
-        t: 'test/{PARENT_FOLDER}/{FILE_NAME}.spec.js,test/*.spec.js',
+        t: 'test/{PARENT}/{FILE_NAME}.spec.js,test/*.spec.js',
         e: 'test/*.spec.js'
       });
       const bitMap = helper.readBitMap();
@@ -623,8 +623,8 @@ describe('bit add command', function () {
       helper.addComponentWithOptions('bar/*.js', {
         i: 'bar/foo',
         m: path.normalize('bar/foo.js'),
-        t: 'test/{PARENT_FOLDER}/{FILE_NAME}.spec.js,bar/foo.spec.js,test/{FILE_NAME}.spec.js',
-        e: 'test/{PARENT_FOLDER}/*.example.*'
+        t: 'test/{PARENT}/{FILE_NAME}.spec.js,bar/foo.spec.js,test/{FILE_NAME}.spec.js',
+        e: 'test/{PARENT}/*.example.*'
       });
       const bitMap = helper.readBitMap();
       const files = bitMap['bar/foo'].files;