+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - implement #613, change `bit install` command to install all packages and link all components
 - change .bit.map.json to .bitmap
+- change {PARENT_FOLDER} variable to {PARENT} in dsl of add
+
 
 ## [0.12.0-dev.26] - 2018-01-15